+++ lib/cuprum/chaining.rb
@@ -9,55 +9,49 @@ module Cuprum
   module Chaining
     # (see Cuprum::BasicCommand#call)
     def call *args, &block
-      call_chained_functions(super)
+      yield_chain(super)
     end # method call
 
-    # Registers a function or block to run after the current function, or after
-    # the last chained function if the current function already has one or more
-    # chained function(s). This creates and modifies a copy of the current
-    # function.
+    # Registers a command or block to run after the current command, or after
+    # the last chained command if the current command already has one or more
+    # chained command(s). This creates and modifies a copy of the current
+    # command.
     #
-    # @param on [Symbol] Sets a condition on when the chained function can run,
-    #   based on the status of the previous function. Valid values are :success,
-    #   :failure, and :always. A value of :success will constrain the function
-    #   to run only if the previous function succeeded. A value of :failure will
-    #   constrain the function to run only if the previous function failed. A
-    #   value of :always will ensure the function is always run, even if the
-    #   function chain has been halted. If no value is given, the function will
-    #   run whether the previous function was a success or a failure, but not if
-    #   the function chain has been halted.
+    # @param on [Symbol] Sets a condition on when the chained command can run,
+    #   based on the status of the previous command. Valid values are :success,
+    #   :failure, and :always. A value of :success will constrain the command
+    #   to run only if the previous command succeeded. A value of :failure will
+    #   constrain the command to run only if the previous command failed. A
+    #   value of :always will ensure the command is always run, even if the
+    #   command chain has been halted. If no value is given, the command will
+    #   run whether the previous command was a success or a failure, but not if
+    #   the command chain has been halted.
     #
-    # @overload chain(function, on: nil)
-    #   The function will be passed the #value of the previous function result
-    #   as its parameter, and the result of the chained function will be
-    #   returned (or passed to the next chained function, if any).
+    # @overload chain(command, on: nil)
+    #   The command will be passed the #value of the previous command result
+    #   as its parameter, and the result of the chained command will be
+    #   returned (or passed to the next chained command, if any).
     #
-    #   @param function [Cuprum::Command] The function to call after the
-    #     current or last chained function.
+    #   @param command [Cuprum::Command] The command to call after the
+    #     current or last chained command.
     #
     # @overload chain(on: :nil, &block)
-    #   The block will be passed the #result of the previous function as its
+    #   The block will be passed the #result of the previous command as its
     #   parameter. If your use case depends on the status of the previous
-    #   function or on any errors generated, use the block form of #chain.
+    #   command or on any errors generated, use the block form of #chain.
     #
     #   If the block returns a Cuprum::Result (or an object responding to #value
     #   and #success?), the block result will be returned (or passed to the next
-    #   chained function, if any). If the block returns any other value
-    #   (including nil), the #result of the previous function will be returned
-    #   or passed to the next function.
+    #   chained command, if any). If the block returns any other value
+    #   (including nil), the #result of the previous command will be returned
+    #   or passed to the next command.
     #
     #   @yieldparam result [Cuprum::Result] The #result of the previous
-    #     function.
+    #     command.
     #
-    # @return [Cuprum::Command] The chained function.
-    def chain function = nil, on: nil, &block
-      clone.tap do |fn|
-        fn.chained_functions <<
-          {
-            :proc => convert_function_or_proc_to_proc(block || function),
-            :on   => on
-          } # end hash
-      end # tap
+    # @return [Cuprum::Command] The chained command.
+    def chain command = nil, on: nil, &block
+      yield_result(:on => on, &chain_command(command || block))
     end # method chain
 
     # Shorthand for function.chain(:on => :failure). Registers a function or
@@ -102,31 +96,33 @@ module Cuprum
       chain(function, :on => :success, &block)
     end # method then
 
+    def yield_result on: nil, &block
+      clone.tap do |fn|
+        fn.chained_procs <<
+          {
+            :proc => block,
+            :on   => on
+          } # end hash
+      end # tap
+    end # method yield_result
+
     protected
 
-    def chained_functions
-      @chained_functions ||= []
-    end # method chained_functions
+    def chained_procs
+      @chained_procs ||= []
+    end # method chained_procs
 
     private
 
-    def call_chained_functions first_result
-      chained_functions.reduce(first_result) do |result, hsh|
-        next result if skip_chained_function?(result, :on => hsh[:on])
-
-        value = hsh.fetch(:proc).call(result)
+    def chain_command command
+      lambda do |result|
+        value = command.call(result)
 
         value_is_result?(value) ? join_result(value) : result
-      end # reduce
-    end # method call_chained_functions
+      end # lambda
+    end # method chain_command
 
-    def convert_function_or_proc_to_proc function_or_proc
-      return function_or_proc if function_or_proc.is_a?(Proc)
-
-      ->(result) { function_or_proc.call(result) }
-    end # method convert_function_or_proc_to_proc
-
-    def skip_chained_function? last_result, on:
+    def skip_chained_proc? last_result, on:
       return false if on == :always
 
       return true if last_result.respond_to?(:halted?) && last_result.halted?
@@ -137,6 +133,20 @@ module Cuprum
       when :failure
         !last_result.failure?
       end # case
-    end # method skip_chained_function?
+    end # method skip_chained_proc?
+
+    def yield_chain first_result
+      chained_procs.reduce(first_result) do |result, hsh|
+        next result if skip_chained_proc?(result, :on => hsh[:on])
+
+        value = hsh.fetch(:proc).call(result)
+
+        if value_is_result?(value)
+          join_result(value)
+        else
+          build_result(value, :errors => build_errors)
+        end # if-else
+      end # reduce
+    end # method yield_chain
   end # module
 end # modue